Fix bug #46863: Agent Queue stats not seen in the report for queues which were connected to some test workload and later deleted

+++ Database/TeleoptiAnalytics/Programmability/03StoredProcedures/mart.etl_bridge_queue_workload_delete.sql
@@ -5,13 +5,24 @@ GO
 -- Description:	delete bridge_queue_workload
 -- =============================================
 CREATE PROCEDURE [mart].[etl_bridge_queue_workload_delete] 
-	   @workload_id int
-      ,@queue_id int
+       @workload_id int
+     , @queue_id int
 AS
 BEGIN
 
-DELETE FROM mart.bridge_queue_workload
-WHERE workload_id=@workload_id AND queue_id=@queue_id
+DECLARE @queueBridgeCount int;
+SELECT @queueBridgeCount = COUNT(*)
+  FROM mart.bridge_queue_workload
+ WHERE queue_id = @queue_id
+
+IF @queueBridgeCount = 1
+  -- If this is the only one record to be deleted, then connect the queue to "Not Defined"
+  UPDATE mart.bridge_queue_workload
+     SET workload_id = -1, skill_id = -1, business_unit_id = -1, datasource_id = -1, update_date = GETDATE()
+   WHERE queue_id = @queue_id
+ELSE IF @queueBridgeCount > 1
+  DELETE FROM mart.bridge_queue_workload
+   WHERE workload_id=@workload_id AND queue_id=@queue_id
 
 END
 GO
